requirements.txt
# Google Generative AI (direct, no LangChain)
google-genai==1.30.0
anyio==4.10.0
websockets==15.0.1

# FastAPI server dependencies
fastapi==0.112.0
uvicorn==0.30.0
python-multipart==0.0.6

# Configuration and data handling
pydantic==2.5.3
python-dotenv==1.0.0

# JSON handling
jsonschema==4.20.0

# Testing
pytest==7.4.3
httpx==0.28.1

# Development tools (optional)
black==23.12.0
flake8==6.1.0
